+++ integration_tests/juniper_tests/src/arc_fields.rs
@@ -1,0 +1,15 @@
+use std::sync::Arc;
+
+struct Query;
+
+#[juniper::graphql_object]
+impl Query {
+    fn ping() -> Arc<bool> {
+        Arc::new(false)
+    }
+}
+
+#[derive(juniper::GraphQLInputObject)]
+struct Ping {
+    expect_result: Arc<bool>,
+}