+++ integration_tests/juniper_tests/src/lib.rs
@@ -1,6 +1,8 @@
 #[cfg(test)]
 mod api;
 #[cfg(test)]
+mod arc_fields;
+#[cfg(test)]
 mod codegen;
 #[cfg(test)]
 mod custom_scalar;