+++ juniper_hyper/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 - Compatibility with the latest `juniper`.
 
+## Breaking Changes
+
+- `juniper_hyper::graphiql` now requires a second parameter for subscriptions
+
 # [[0.5.2] 2019-12-16](https://github.com/graphql-rust/juniper/releases/tag/juniper_hyper-0.5.2)
 
 - Compatibility with the latest `juniper`.