+++ juniper_hyper/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 - Compatibility with the latest `juniper`.
 
+# [[0.6.0] 2020-12-09](https://github.com/graphql-rust/juniper/releases/tag/juniper_hyper-0.6.0)
+
+- Compatibility with the latest `juniper`.
+
 ## Breaking Changes
 
 - `juniper_hyper::graphiql` now requires a second parameter for subscriptions.